Created new prefabs and reworked some levels.
Added value to the input buffer inside the player prefab. Changed Lv11. Created new Lv9 and added it to the BuildSettings (instead of the old Lv9, now renamed Lv9OLD). Created prefab for moving obstacle (now with a different color to better identify it) and a prefab for the waypoints used for it.