Add CanPressButton to SecurityBox
Plant can't press buttons if they're just spawned. Since the
collision happens between SwitchButton and the plant SecurityBox
child, the latter gives the information to SwitchButton via
CanPressButton method, which checks if enough time has passed
since the instance initialization. Check SecurityBox code for
more info.

Amount of time needed to make the plant able to press button can
be altered by changing the timeBeforeInteraction variable in
SecurityBox script.

TODO: refactor both sides (SwitchButton and SecurityBox) of this
feature implementation when we have some animation made for the
plant. It would be best to link the functionality to the animation
state instead of some arbitrary float value. Search for TEMP tags
in the code to see what must be refactored.