Slight changes to Lv24, 25, 26 WIP and 27 WIP. Created Lv26 and 27.
Previous Lv26 and 27 are nore renamed 26 WIP and 27 WIP as they contain tests for other levels. Created Lv26 and 27. These levels are complete and can be used. Also added them to the editor settings.
Created LvBase to use it as a fundation when creating new levels. Implemented test tile 1700x800.