Add DestroyParticles, Water tag, bubbles to camera
DestroyParticles destroy particles from a ParticleSystem when they
reach a certain position relative to the Limit object position,
plus/minus a given offset.
A positive offset always allow particles to go "beyond" the given
Limit, so if "above" is chosen the offset makes the particles go
above it by a certain amount, etc.
If no Custom Limit is given, Water is chosen as Limit by default.

Water must be referenced by tag "Water" from now on.

The bubble emitter is a child of MainCamera and moves with it.