New stages. Changes to teleport related prefabs.
Added component relative to the teleport in the prefab of the player.  Increased teleportDistance in the TargetPlaceholder and the TeleportSoil. Enabled collider in the TeleportPlant teleport and activated isTrigger (to make it so that the water starts when a teleport plant is created).
Removed a cube in Level_1. Slight change to LvBase to avoid problems with the camera in the Window version of the game. Commented a line in the TargetManager script.
Created Level_31, 32 and 33 (these will be the first levels of the World 2). Added them to the BuildSettings.